% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{enrichMetrics}
\alias{enrichMetrics}
\title{An internal function}
\usage{
enrichMetrics(title.results, title.page, title.id)
}
\description{
Used to add some extra information to a movieMetrics object.
  This function should only be called within fetchMovie().
}
\seealso{
fetchMovie()
}
