% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{fetchMoviesList}
\alias{fetchMoviesList}
\title{Core function}
\usage{
fetchMoviesList(titles, moviesDB)
}
\arguments{
\item{title}{a character vector of any length}
}
\value{
a movieMetricsList object

a movieMetricsList object
}
\description{
Used to retrieve data about a list of movies using MOJO database
}
\examples{
titles <- searchTitles('star wars')
fetchMoviesList(titles)

}
\seealso{
fetchMovie() fetchMoviesList()
}
\keyword{movieMetricsList}
