% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{fixMetrics}
\alias{fixMetrics}
\title{An internal function}
\usage{
fixMetrics(title.results)
}
\description{
Used to transform the data retrieved from MOJO intro an organized dataframe.
  This function should only be called within fetchMovie().
}
\seealso{
fetchMovie()
}
