% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{data}
\name{moviesDB}
\alias{moviesDB}
\title{moviesDB object}
\format{An object of class \code{moviesDB} (inherits from \code{list}) of length 2.}
\usage{
moviesDB
}
\description{
An object of the moviesDB class.
Movies database generated from The Numbers.
It was generated in April 2019 and contains 5,200 movies at that time. 
To update the database with all the existing movies, please see examples.
}
\examples{
moviesDB <- availableTitles(n.pages = 52, verbose = TRUE)

}
\keyword{datasets}
