% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{searchTitle}
\alias{searchTitle}
\title{A useful search function}
\usage{
searchTitle(pattern, moviesDB, exact.match = FALSE,
  return.list = FALSE, return.url = FALSE)
}
\arguments{
\item{pattern}{a character vector of length 1}

\item{moviesDB}{a list of length 2, the first element containing movies 
names and the second associated URLs.}

\item{exact.match}{logical. Whether to perform an exact match}

\item{return.list}{logical. Whether to return the list of matched titles}

\item{return.url}{logical. Whether to return URLs instead of matched titles 
names (cannot be used with `return.list = T`)}
}
\value{
a vector of matching title or matching URLs
}
\description{
Used to find all titles in MOJO database matching a specific pattern
}
\examples{
searchTitle('star wars', moviesDB)

}
\seealso{
fetchMovie()
}
