% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{searchTitles}
\alias{searchTitles}
\title{A useful search function}
\usage{
searchTitles(pattern)
}
\arguments{
\item{pattern}{a character vector of length 1}
}
\value{
a vector of matching titles
}
\description{
Used to find all titles in MOJO database matching a specific pattern
}
\examples{
searchTitles('star wars')

}
\seealso{
fetchMovie()
}
\keyword{search}
\keyword{title}
