% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{availableTitles}
\alias{availableTitles}
\title{A core function}
\usage{
availableTitles(n.pages = 500, verbose = T)
}
\arguments{
\item{n.pages}{an integer. The number of pages to scan}
}
\value{
a list of length 2. The first element of the list is a vector 
  of available titles. The second element contains the corresponding urls.
}
\description{
Used to find all titles available in the first n.pages of The Numbers database
}
\examples{
searchTitles('star wars')

}
\seealso{
fetchMovie() searchTitle()
}
\keyword{searchTitle}
