% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{fetchMovie}
\alias{fetchMovie}
\title{Core function}
\usage{
fetchMovie(title)
}
\arguments{
\item{title}{a character vector of length 1}
}
\value{
a movieMetrics object
}
\description{
Used to retrieve data about a movie using MOJO database
}
\examples{
fetchMovie('Star Wars')

}
\seealso{
fetchMovie() fetchMoviesList()
}
\keyword{movieMetrics}
