% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetchMovie}
\alias{fetchMovie}
\title{Core function}
\usage{
fetchMovie(title, moviesDB)
}
\arguments{
\item{title}{a character vector of length 1}

\item{moviesDB}{a list of length 2, the first element containing movies 
names and the second associated URLs.}
}
\value{
a movieMetrics object

a movieMetrics object
}
\description{
Used to retrieve data about a movie using MOJO database
}
\examples{
fetchMovie('The Greatest Showman', moviesDB)

}
\seealso{
fetchMovie() fetchMoviesList()
}
\keyword{movieMetrics}
