% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{fetchMoviesList}
\alias{fetchMoviesList}
\title{Core function}
\usage{
fetchMoviesList(titles, moviesDB)
}
\arguments{
\item{titles}{a character vector of any length}

\item{moviesDB}{a list of length 2, the first element containing movies 
names and the second associated URLs.}
}
\value{
a movieMetricsList object

a movieMetricsList object
}
\description{
Used to retrieve data about a list of movies using MOJO database
}
\examples{
moviesDB <- availableTitles(n.pages = 20, verbose = TRUE)
starwarsmovies <- searchTitle('star wars', moviesDB, return.list = TRUE)
starwarsmovies <- fetchMoviesList(starwarsmovies, moviesDB)

}
\seealso{
fetchMovie() fetchMoviesList()
}
\keyword{movieMetricsList}
